+++ common/src/main/java/com/linecorp/line/auth/fido/fido2/common/mdsv3/metadata/PatternAccuracyDescriptor.java
@@ -9,11 +9,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class PatternAccuracyDescriptor {
-    private long minComplexity; //unsigned long
+    private BigInteger minComplexity; //unsigned long
     private Integer maxRetries;
     private Integer blockSlowdown;
 }
